requirements_hf.txt
streamlit==1.32.0
sentence-transformers==2.5.1
transformers==4.38.2
torch==2.2.1
google-generativeai==0.4.1
chromadb==0.4.24
spacy==3.7.4
scispacy==0.5.4
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz
pypdf==4.1.0
python-dotenv==1.0.1
